requirements.txt
openai>=1.43.0,<1.44.0
agentops>=0.3.10,<0.4.0
python-dotenv>=1.0.1,<2.0.0